chore: moves pr loader to its own file.
