+++ glide.lock
@@ -1,23 +1,35 @@
-hash: 2dc53922a6948058f6a4520a20e9f4b69c96f61846d2c02f4fd32a29b48e81d8
-updated: 2018-04-08T22:56:19.475386739+02:00
+hash: 84a14ebad800adab02ba6e313744eaecf634b1ad07bad12354d4eecf2abb78fa
+updated: 2018-04-09T15:30:48.995447157+02:00
 imports:
 - name: github.com/bazelbuild/buildtools
-  version: a8cd34f034f2ae1e206eec896cf12d38a0cb26fb
+  version: c98ff0c6395f09b1942e6f7c42bf3ec15e3b9ca7
   subpackages:
   - build
   - tables
 - name: github.com/beorn7/perks
-  version: 3a771d992973f24aa725d07868b467d1ddfceafb
+  version: 3ac7bf7a47d159a033b107610db8a1b6575507a4
   subpackages:
   - quantile
 - name: github.com/certifi/gocertifi
   version: deb3ae2ef2610fde3330947281941c562861188b
+- name: github.com/emicklei/go-restful
+  version: ff4f55a206334ef123e4f79bbf348980da81ca46
+  subpackages:
+  - log
 - name: github.com/evalphobia/logrus_sentry
   version: 57846a82817615f185b10cc40de8dc60c1ca5ae1
 - name: github.com/getsentry/raven-go
   version: 563b81fc02b75d664e54da31f787c2cc2186780b
 - name: github.com/ghodss/yaml
   version: 73d445a93680fa1a78ae23a5839bad48f32ba1ee
+- name: github.com/go-openapi/jsonpointer
+  version: 46af16f9f7b149af66e5d1bd010e3574dc06de98
+- name: github.com/go-openapi/jsonreference
+  version: 13c6e3589ad90f49bd3e3bbe2c2cb3d7a4142272
+- name: github.com/go-openapi/spec
+  version: 6aced65f8501fe1217321abf0749d354824ba2ff
+- name: github.com/go-openapi/swag
+  version: 1d0bd113de87027671077d3c71eb3ac5d7dbba72
 - name: github.com/gogo/protobuf
   version: c0656edd0d9eab7c66d1eb0c568f9039345796f7
   subpackages:
@@ -26,9 +38,9 @@ imports:
 - name: github.com/golang/glog
   version: 44145f04b68cf362d9c4df2182967c2275eaefed
 - name: github.com/golang/lint
-  version: 85993ffd0a6cd043291f3f63d45d656d97b165bd
+  version: c7bacac2b21ca01afa1dee0acf64df3ce047c28f
 - name: github.com/golang/protobuf
-  version: e09c5db296004fbe3f74490e84dcd62c3c5ddb1b
+  version: 8ee79997227bf9b34611aee7946ae64735e6fd93
   subpackages:
   - proto
 - name: github.com/google/go-github
@@ -36,7 +48,7 @@ imports:
   subpackages:
   - github
 - name: github.com/google/go-querystring
-  version: 53e6ce116135b80d037921a7fdd5138cf32d7a8a
+  version: 9235644dd9e52eeae6fa48efd539fdc351a0af53
   subpackages:
   - query
 - name: github.com/google/gofuzz
@@ -46,9 +58,15 @@ imports:
 - name: github.com/gorilla/securecookie
   version: e59506cc896acb7f7bf732d4fdf5e25f7ccd8983
 - name: github.com/gorilla/sessions
-  version: a2f2a3de9a4a575047f73e3e36bc85ecc3546391
+  version: 7087b4d669d1bc3da42fb4e2eda73ae139a24439
+- name: github.com/mailru/easyjson
+  version: d5b7844b561a7bc640052f1b935f7b800330d7e0
+  subpackages:
+  - buffer
+  - jlexer
+  - jwriter
 - name: github.com/matttproud/golang_protobuf_extensions
-  version: c12348ce28de40eed0136aa2b644d0ee0650e56c
+  version: fc2b8d3a73c4867e51861bbdd5ae3c1f0869dd6a
   subpackages:
   - pbutil
 - name: github.com/onsi/ginkgo
@@ -87,49 +105,48 @@ imports:
   - matchers/support/goraph/util
   - types
 - name: github.com/pkg/errors
-  version: 816c9085562cd7ee03e7f8188a1cfd942858cded
+  version: 645ef00459ed84a119197bfb8d8205042c6df63d
 - name: github.com/prometheus/client_golang
-  version: f504d69affe11ec1ccb2e5948127f86878c9fd57
+  version: e51041b3fa41cece0dca035740ba6411905be473
   subpackages:
   - prometheus
 - name: github.com/prometheus/client_model
-  version: 99fa1f4be8e564e8a6b613da7fa6f46c9edafc6c
+  version: fa8ad6fec33561be4280a8f0514318c79d7f6cb6
   subpackages:
   - go
 - name: github.com/prometheus/common
-  version: 38c53a9f4bfcd932d1b00bfc65e256a7fba6b37a
+  version: 2e54d0b93cba2fd133edc32211dcc32c06ef72ca
   subpackages:
   - expfmt
   - internal/bitbucket.org/ww/goautoneg
   - model
 - name: github.com/prometheus/procfs
-  version: 780932d4fbbe0e69b84c34c20f5c8d0981e109ea
-  subpackages:
-  - internal/util
-  - nfs
-  - xfs
+  version: 454a56f35412459b5e684fd5ec0f9211b94f002a
+- name: github.com/PuerkitoBio/purell
+  version: 8a290539e2e8629dbc4e6bad948158f790ec31f4
+- name: github.com/PuerkitoBio/urlesc
+  version: 5bd2802263f21d8788851d5305584c82a5c75d7e
 - name: github.com/satori/go.uuid
   version: 879c5887cd475cd7864858769793b2ceb0d44feb
 - name: github.com/shurcooL/githubql
-  version: d8297a7d5a840a2105de3f27796a6876d03b5da1
+  version: a82ea8c70b7fd207aeb72ccb10575eed252f2f68
 - name: github.com/shurcooL/go
   version: 364c5ae8518b51f5fda954762864d6f4d5c30c89
   subpackages:
   - ctxhttp
 - name: github.com/shurcooL/graphql
-  version: 3d276b9dcc6b1e0adf19557a8de5cb8632c07697
+  version: d0549edd16dceb6939e538fdb1b4f2ec7ee816cc
   subpackages:
   - ident
   - internal/jsonutil
 - name: github.com/sirupsen/logrus
   version: 4b6ea7319e214d98c938f12692336f7ca9348d6b
 - name: github.com/spf13/pflag
-  version: 4c012f6dcd9546820e378d0bdda4d8fc772cdfea
+  version: 9ff6c6923cfffbcd502984b8e0c80539a94968b7
 - name: golang.org/x/net
   version: 1c05540f6879653db88113bc4a2b70aec4bd491f
   subpackages:
   - context
-  - context/ctxhttp
   - html
   - html/atom
   - html/charset
@@ -138,16 +155,17 @@ imports:
   - idna
   - lex/httplex
 - name: golang.org/x/oauth2
-  version: 921ae394b9430ed4fb549668d7b087601bd60a81
+  version: a6bd8cefa1811bd24b86f8902872e4e8225f74c4
   subpackages:
   - internal
 - name: golang.org/x/sys
-  version: 1d206c9fa8975fb4cf00df1dc8bf3283dc24ba0e
+  version: 8f0908ab3b2457e2e15403d3697c9ef5cb4b57a9
   subpackages:
   - unix
 - name: golang.org/x/text
   version: b19bf474d317b857955b12035d2c5acb57ce8b01
   subpackages:
+  - cases
   - encoding
   - encoding/charmap
   - encoding/htmlindex
@@ -158,22 +176,23 @@ imports:
   - encoding/simplifiedchinese
   - encoding/traditionalchinese
   - encoding/unicode
+  - internal
   - internal/tag
   - internal/utf8internal
   - language
   - runes
   - secure/bidirule
+  - secure/precis
   - transform
   - unicode/bidi
   - unicode/norm
+  - width
 - name: golang.org/x/tools
-  version: ac136b6c2db7c4d43955e4bc7174db36dc0539c0
+  version: a888bfdffa4526cc6987572bca9a2c6b7758290f
   subpackages:
-  - go/ast/astutil
-  - go/gcexportdata
   - go/gcimporter15
 - name: google.golang.org/appengine
-  version: 0a24098c0ec68416ec050f567f75df563d6b231e
+  version: 12d5545dc1cfa6047a286d5e853841b6471f4c19
   subpackages:
   - internal
   - internal/base
@@ -189,13 +208,13 @@ imports:
 - name: gopkg.in/robfig/cron.v2
   version: be2e0b0deed5a68ffee390b4583a13aff8321535
 - name: gopkg.in/yaml.v2
-  version: 5420a8b6744d3b0345ab293f6fcba19c978f1183
+  version: a5b47d31c556af34a302ce5d659e6fea44d90de0
 - name: k8s.io/api
-  version: c4ebf33f93081f8c7602d29a59944e4a9f684ac2
+  version: 4df58c811fe2e65feb879227b2b245e4dc26e7ad
   subpackages:
   - core/v1
 - name: k8s.io/apimachinery
-  version: 0ed326127d3068118ef128c76673dad6005736ba
+  version: 019ae5ada31de202164b118aee88ee2d14075c31
   subpackages:
   - pkg/api/resource
   - pkg/apis/meta/v1
@@ -209,7 +228,6 @@ imports:
   - pkg/types
   - pkg/util/errors
   - pkg/util/intstr
-  - pkg/util/json
   - pkg/util/net
   - pkg/util/runtime
   - pkg/util/sets
@@ -218,6 +236,10 @@ imports:
   - pkg/util/wait
   - pkg/watch
   - third_party/forked/golang/reflect
+- name: k8s.io/kube-openapi
+  version: 868f2f29720b192240e18284659231b440f9cda5
+  subpackages:
+  - pkg/common
 - name: k8s.io/test-infra
   version: 2f3ad698ffcff53f881d3292a1b3da894bf1d5f7
   subpackages:
