+++ pkg/github/client_test.go
@@ -3,7 +3,6 @@ package github_test
 import (
 	"net/http"
 
-	"github.com/arquillian/ike-prow-plugins/pkg/github"
 	. "github.com/arquillian/ike-prow-plugins/pkg/internal/test"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -64,53 +63,6 @@ var _ = Describe("Client features", func() {
 		})
 	})
 
-	Context("Lazy loading feature", func() {
-
-		BeforeEach(func() {
-			gock.Off()
-		})
-
-		It("should load pull request when load() method is called", func() {
-			// given
-			counter := 0
-			gock.New("https://api.github.com").
-				Get("/repos/owner/repo/pulls/123").
-				SetMatcher(createCounterMather(&counter)).
-				Reply(200).
-				BodyString(`{"title":"Loaded PR"}`)
-			loader := &github.PullRequestLoader{Client: client, RepoOwner: "owner", RepoName: "repo", Number: 123}
-			Expect(counter).To(Equal(0))
-
-			// when
-			pullRequest, err := loader.Load()
-
-			// then
-			Ω(err).ShouldNot(HaveOccurred())
-			Expect(counter).To(Equal(1))
-			Expect(*pullRequest.Title).To(Equal("Loaded PR"))
-		})
-
-		It("should load pull request only once", func() {
-			// given
-			counter := 0
-			gock.New("https://api.github.com").
-				Get("/repos/owner/repo/pulls/123").
-				SetMatcher(createCounterMather(&counter)).
-				Persist().
-				Reply(200).
-				BodyString(`{"title":"Loaded PR"}`)
-			loader := &github.PullRequestLoader{Client: client, RepoOwner: "owner", RepoName: "repo", Number: 123}
-			loader.Load()
-
-			// when
-			pullRequest, err := loader.Load()
-
-			// then
-			Ω(err).ShouldNot(HaveOccurred())
-			Expect(counter).To(Equal(1))
-			Expect(*pullRequest.Title).To(Equal("Loaded PR"))
-		})
-	})
 })
 
 func createCounterMather(counter *int) gock.Matcher {
