+++ pkg/plugin/test-keeper/plugin/comment_cmd.go
@@ -5,8 +5,8 @@ import (
 
 	is "github.com/arquillian/ike-prow-plugins/pkg/command"
 	"github.com/arquillian/ike-prow-plugins/pkg/github"
-	gogh "github.com/google/go-github/github"
 	"github.com/arquillian/ike-prow-plugins/pkg/log"
+	gogh "github.com/google/go-github/github"
 )
 
 // SkipComment is used as a command to bypass test presence validation
